Update email subjects and contents, add new subscribers, and integrate the new DDOS prevention workflow
In this commit, we updated the contents and subjects of several emails in the application, including an invitation, event reminder, and others. We added new subscribers to the loadSubscribers.ts file. Additionally, we integrated a new workflow, "Novarian-DDOS", which includes sending a "ddos" email to select subscribers. Environment variables required for the application to run have been made available in a new .env file.