Updated various scripts and added new animations to dashboard
In this commit, the DDoS script was deleted, loadSubscribers.ts was modified to add subscribers in bulk, and the subject string in client.ts was corrected. A new file startEvent.ts was created which exports an asynchronous function. Changes were made in two other scripts, ddos.ts and page.tsx, to make the application more interactive and responsive. 'novu' and 'framer-motion' in package.json and package-lock.json were updated and other related dependencies were also added.