Update app components, configuration, and dependencies
This commit deletes 'evervault-card' and 'ddos' components, introduces 'event' page, and updates the 'Header' and 'app' components. It also enhances the configuration in 'next.config.mjs' to include new settings and adjusts the 'package.json' to update the app version and several dependencies. Additionally, an endpoint URL in 'commence.tsx' was changed.